Bump CI images to their latest two versions

+++ tests/encode/main.c
@@ -10,6 +10,8 @@
 
 #include "nanocbor/nanocbor.h"
 
+/* NOLINTBEGIN(cppcoreguidelines-avoid-magic-numbers,readability-magic-numbers)
+ */
 static void _encode(nanocbor_encoder_t *enc)
 {
     nanocbor_fmt_array_indefinite(enc);
@@ -24,12 +26,13 @@ static void _encode(nanocbor_encoder_t *enc)
     nanocbor_fmt_int(enc, 500);
     nanocbor_fmt_int(enc, -500);
     nanocbor_put_tstr(enc, "this is a long string");
-    nanocbor_fmt_float(enc, 0.34);
+    nanocbor_fmt_float(enc, (float)0.34);
     nanocbor_put_bstr(enc, (uint8_t *)"bytez", sizeof("bytez"));
     nanocbor_fmt_null(enc);
     nanocbor_fmt_decimal_frac(enc, -2, 27315);
     nanocbor_fmt_end_indefinite(enc);
 }
+/* NOLINTEND(cppcoreguidelines-avoid-magic-numbers,readability-magic-numbers) */
 
 int main(void)
 {
